docker/range-des-weak/requirements.txt
Flask==3.0.0
pyyaml==6.0.1
pycryptodome==3.19.0
